+++ .vscode/settings.json
@@ -23,5 +23,9 @@
         "github.copilot",
         "github.copilot-chat"
     ],
-    "ruff.trace.server": "messages"
+    "ruff.trace.server": "messages",
+    "python.sortImports.args": ["--profile", "black"],
+    "editor.codeActionsOnSave": {
+      "source.organizeImports": true
+    }
 }
