+++ deep_quoridor/src/play.py
@@ -2,7 +2,7 @@ import argparse
 
 from agents import AgentRegistry
 from arena import Arena
-from arena_yaml_recorder import ArenaYAMLRecorder
+from plugins.arena_yaml_recorder import ArenaYAMLRecorder
 from renderers import Renderer
 from utils.misc import set_deterministic
