+++ deep_quoridor/src/plugins/__init__.py
@@ -1,0 +1,2 @@
+from plugins.arena_yaml_recorder import ArenaYAMLRecorder  # noqa: E402, F401
+from plugins.save_model import SaveModelEveryNEpisodesPlugin  # noqa: E402, F401
