+++ deep_quoridor/src/renderers/__init__.py
@@ -27,5 +27,6 @@ from renderers.curses_board import CursesBoardRenderer  # noqa: E402, F401
 from renderers.match_results import MatchResultsRenderer  # noqa: E402, F401
 from renderers.none import NoneRenderer  # noqa: E402, F401
 from renderers.progress_bar import ProgressBarRenderer  # noqa: E402, F401
-from renderers.text_board import TextBoardRenderer  # noqa: E402, F401
 from renderers.pygame import PygameRenderer  # noqa: E402, F401
+from renderers.text_board import TextBoardRenderer  # noqa: E402, F401
+from renderers.training_status import TrainingStatusRenderer  # noqa: E402, F401
