+++ deep_quoridor/src/train.py
@@ -4,85 +4,9 @@ import os
 import torch
 from agents import DExpAgent, GreedyAgent, SimpleAgent
 from agents.dexp import DExpPlayParams
-from agents.flat_dqn import AbstractTrainableAgent
 from arena import Arena
-from arena_utils import ArenaPlugin
-from renderers import Renderer
-from utils.misc import set_deterministic
-
-
-class TrainingStatusRenderer(Renderer):
-    def __init__(self, update_every: int, total_episodes: int, agents: list[AbstractTrainableAgent]):
-        self.agents = agents
-        self.update_every = update_every
-        self.total_episodes = total_episodes
-        self.episode_count = 0
-
-    def start_game(self, game, agent1, agent2):
-        self.step = 0
-        self.player1 = agent1.name()
-        self.player2 = agent2.name()
-
-    def end_game(self, game, result):
-        if self.episode_count % self.update_every == 0:
-            for agent in self.agents:
-                agent_name = agent.name()
-                avg_reward = (
-                    sum(agent.episodes_rewards[-1 * self.update_every :])
-                    / min(self.update_every, len(agent.episodes_rewards))
-                    if agent.episodes_rewards
-                    else 0.0
-                )
-                avg_loss = (
-                    sum(agent.train_call_losses[-1 * self.update_every :])
-                    / min(self.update_every, len(agent.train_call_losses))
-                    if agent.train_call_losses
-                    else 0.0
-                )
-                print(
-                    f"{agent_name} Episode {self.episode_count + 1}/{self.total_episodes}, Steps: {self.step}, Avg Reward: {avg_reward:.2f}, "
-                    f"Avg Loss: {avg_loss:.4f}, Epsilon: {agent.epsilon:.4f}"
-                )
-        self.episode_count += 1
-        return
-
-    def action(self, game, step, agent, action):
-        self.step += 1
-        if self.step == 1000:
-            print("board")
-            print(game.render())
-            print(f"player0: {self.player1}")
-            print(game.observe("player_0"))
-            print(f"player1:  {self.player2}")
-            print(game.observe("player_1"))
-
-
-class SaveModelEveryNEpisodesPlugin(ArenaPlugin):
-    def __init__(
-        self, update_every: int, path: str, board_size: int, max_walls: int, agents: list[AbstractTrainableAgent]
-    ):
-        self.agents = agents
-        self.update_every = update_every
-        self.path = path
-        self.episode_count = 0
-        self.board_size = board_size
-        self.max_walls = max_walls
-
-    def end_game(self, game, result):
-        if self.episode_count % self.update_every == 0 and self.episode_count > 0:
-            self._save_models(f"_episode_{self.episode_count}")
-        self.episode_count += 1
-
-    def end_arena(self, game, results):
-        self._save_models("final")
-
-    def _save_models(self, suffix: str):
-        for agent in self.agents:
-            agent_name = agent.name()
-            filename = agent.resolve_filename(suffix)
-            save_file = os.path.join(self.path, filename)
-            agent.save_model(save_file)
-            print(f"{agent_name} Model saved to {save_file}")
+from plugins import SaveModelEveryNEpisodesPlugin
+from renderers import TrainingStatusRenderer
 
 
 def train_dqn(
